+++ composer.json
@@ -35,7 +35,8 @@
         "symfony/validator": "^3.3",
         "symfony/config": "^3.3",
         "symfony/doctrine-bridge": "^3.3",
-        "symfony/translation": "^3.3"
+        "symfony/translation": "^3.3",
+        "league/commonmark": "^0.15.4"
     },
     "autoload": {
         "psr-4": {
