+++ src/Application/Provider/UserProvider.php
@@ -40,7 +40,7 @@ class UserProvider implements UserProviderInterface
             throw new UsernameNotFoundException(sprintf('Username "%s" does not exist.', $username));
         }
 
-        return new User($user['phone'], $user['password'], ['ROLE_USER'], true, true, true, true);
+        return new User($user['phone'], $user['password'], ['ROLE_USER', 'ROLE_ADMIN'], true, true, true, true);
     }
 
     /**
