+++ templates/admin/publishers/index.twig
@@ -1,22 +1,17 @@
 {% extends "admin/base.twig" %}
 {% block title %}Publisher Details{% endblock %}
 {% block content %}
-    <div class="w-section section">
-        <div class="content-mobile-number">
-            <img class="carty" src="/images/happy-13.svg">
-            <div class="login-div" id="container">
-                <img class="leaf" src="/images/leaf-11.svg">
-                <div class="sign-in-text">Publishers</div>
-                {% for message in app.session.flashbag.get('info') %}
-                    <div class="flash-info alert alert-success">{{ message }}</div>
-                {% endfor %}
-                {% include 'admin/publishers/search_form.twig' %}
-                <div class="data" id="results"></div>
-                  {% form_theme form 'bootstrap_3_horizontal_layout.html.twig' %}
-                  {{ form_start(form) }}
-                  {{ form_widget(form) }}
-                  {{ form_end(form) }}
-            </div>
-        </div>
+    <div class="container-fluid">
+        <img class="leaf" src="/images/leaf-11.svg">
+        <div class="sign-in-text">Publishers</div>
+      {% for message in app.session.flashbag.get('info') %}
+          <div class="flash-info alert alert-success">{{ message }}</div>
+      {% endfor %}
+      {% include 'admin/publishers/search_form.twig' %}
+        <div class="data" id="results"></div>
+      {% form_theme form 'bootstrap_3_horizontal_layout.html.twig' %}
+      {{ form_start(form) }}
+      {{ form_widget(form) }}
+      {{ form_end(form) }}
     </div>
 {% endblock %}
