+++ templates/admin/publishers/search_form.twig
@@ -1,4 +1,4 @@
-<form method="post" action="/admin/publishers/search">
+<form method="post" action="{{ path('admin/publishers/search') }}">
   <div class="textSearch">
     <input type="hidden" name="action" value="search">
     <input type="search" class="form-control" id="name" placeholder="Search a publisher" name="name">
