+++ templates/index.twig
@@ -4,9 +4,9 @@
 <div class="w-nav nav-global" data-collapse="all" data-animation="default" data-duration="400" data-contain="1">
     <div class="w-container">
         <nav class="w-nav-menu menu-background" role="navigation">
-            <a class="w-nav-link nav-link" href="/placements">Record placements</a>
-            <a target="_blank" class="w-nav-link nav-link" href="/map">Maps</a>
-            <a class="w-nav-link nav-link" href="/logout.php">Sign out</a>
+            <a class="w-nav-link nav-link" href="{{ path('placements') }}">Record placements</a>
+            <a target="_blank" class="w-nav-link nav-link" href="{{ path('locations') }}">Maps</a>
+            <a class="w-nav-link nav-link" href="{{ path('placements') }}">Sign out</a>
         </nav>
         <div class="w-nav-button menu-bt">
             <div class="w-icon-nav-menu"></div>
