--- templates/reports.twig
@@ -1,21 +1,0 @@
-{% extends "base.twig" %}
-{% block title %}Metropolitan Witnessing{% endblock %}
-{% block content %}
-    <div class="w-section section">
-        <div class="content-mobile-number"><img class="carty" src="../images/happy-13.svg">
-            <div class="login-div"><img class="leaf" src="../images/leaf-11.svg">
-                <div class="sign-in-text"></div>
-                <div class="w-form">
-                    <form method="post" action="/reports" id="brothersReport" name="brothers-report" >
-                        <input class="w-button submit-mobile-number sign-in" type="submit" name="action" value="List Brothers">
-                    </form>
-                </div>
-                <div class="w-form">
-                    <form method="post" action="/reports" id="inviteesReport" name="invitees-report" >
-                        <input class="w-button submit-mobile-number sign-in" type="submit" name="action" value="List Invitees">
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-{% endblock %}
